move logo stuff into separate file
